API: removed deprecated function pw_distances_from_table
